Added partial objects
Added include_paths() and exclude_paths() to View, to allow retrieving
partial documents in searches.  The UID of a partial document is
marked as is_partial, which prevents the document being saved.

Partial objects are intended as light-weight objects purely for displaying
results, where only a few fields are required.

Closes #6